@@ -6,45 +6,46 @@ repository: https://github.com/GetStream/stream-chat-flutter
 issue_tracker: https://github.com/GetStream/stream-chat-flutter/issues
 
 environment:
-  sdk: ">=3.1.0 <4.0.0"
-  flutter: ">=3.13.0"
+  sdk: ">=3.3.0 <4.0.0"
+  flutter: ">=3.19.0"
 
 dependencies:
-  cached_network_image: ^3.2.3
-  chewie: ^1.7.0
+  cached_network_image: ^3.3.1
+  chewie: ^1.8.1
   collection: ^1.17.2
   contextmenu: ^3.0.0
-  dart_vlc: ^0.4.0
-  desktop_drop: ^0.4.1
-  diacritic: ^0.1.4
-  dio: ^5.3.2
+  desktop_drop: ^0.4.4
+  diacritic: ^0.1.5
+  dio: ^5.4.3+1
   ezanimation: ^0.6.0
-  file_picker: ">=5.3.3 <7.0.0"
-  file_selector: ^1.0.0
+  file_picker: ^8.0.5
+  file_selector: ^1.0.3
   flutter:
     sdk: flutter
-  flutter_markdown: ^0.6.17+1
+  flutter_markdown: ^0.7.2+1
   flutter_portal: ^1.1.4
-  flutter_svg: ^2.0.7
+  flutter_svg: ^2.0.10+1
   http_parser: ^4.0.2
   image_gallery_saver: ^2.0.3
-  image_picker: ^1.0.2
-  image_size_getter: ^2.1.2
+  image_picker: ^1.1.2
+  image_size_getter: ^2.1.3
   jiffy: ^6.2.1
-  just_audio: ^0.9.37
+  just_audio: ^0.9.38
   lottie: ">=2.6.0 <4.0.0"
+  media_kit: ^1.1.10+1
+  media_kit_video: ^1.2.4
   meta: ^1.9.1
-  path_provider: ^2.1.0
-  photo_manager: ^3.0.0
-  photo_view: ^0.14.0
+  path_provider: ^2.1.3
+  photo_manager: ^3.2.0
+  photo_view: ^0.15.0
   rxdart: ^0.27.7
-  share_plus: ">=7.1.0 <9.0.0"
+  share_plus: ">=8.0.2 <10.0.0"
   shimmer: ^3.0.0
   stream_chat_flutter_core: ^7.2.2
-  synchronized: ^3.1.0
+  synchronized: ^3.1.0+1
   thumblr: ^0.0.4
-  url_launcher: ^6.1.12
-  video_player: ^2.7.0
+  url_launcher: ^6.3.0
+  video_player: ^2.8.7
   video_thumbnail: ^0.5.3
 
 flutter:
