@@ -6,18 +6,18 @@ repository: https://github.com/GetStream/stream-chat-flutter
 issue_tracker: https://github.com/GetStream/stream-chat-flutter/issues
 
 environment:
-  sdk: ">=3.1.0 <4.0.0"
-  flutter: ">=3.13.0"
+  sdk: ">=3.3.0 <4.0.0"
+  flutter: ">=3.19.0"
 
 dependencies:
-  drift: ^2.11.0
+  drift: ^2.18.0
   flutter:
     sdk: flutter
   logging: ^1.2.0
   meta: ^1.9.1
   path: ^1.8.3
-  path_provider: ^2.1.0
-  sqlite3_flutter_libs: ^0.5.15
+  path_provider: ^2.1.3
+  sqlite3_flutter_libs: ^0.5.24
   stream_chat: ^7.2.2
 
 dev_dependencies:
