@@ -69,8 +69,8 @@ void main() {
             client: client,
             child: StreamChannel(
               channel: channel,
-              child: WillPopScope(
-                onWillPop: () async => false,
+              child: PopScope(
+                onPopInvoked: (bool didPop) async => false,
                 child: Scaffold(
                   appBar: StreamGalleryHeader(
                     attachment: MockAttachment(),
@@ -94,8 +94,8 @@ void main() {
           client: client,
           child: StreamChannel(
             channel: channel,
-            child: WillPopScope(
-              onWillPop: () async => false,
+            child: PopScope(
+              onPopInvoked: (bool didPop) async => false,
               child: Scaffold(
                 appBar: StreamGalleryHeader(
                   userName: 'User',
